Maintain source for pure python WordHunt.
Adjust alias in the init.
