+++ app_wordhunt/__init__.py
@@ -1,7 +1,8 @@
 """WordHunt board solving app."""
 import utils
 
-from . import wordhunt
+import pyximport; pyximport.install()
+from . import cy_wordhunt as wordhunt
 
 
 APP_HELP = "Solve a WordHunt board. Content is the board - top to bottom, left to right."
